requirements.txt
requests
begins
path.py
tqdm
